Move the tests for winner-recording buttons into their own module
